requirements.txt
pyqt6
requests
pyserial